Migrated Code & Initial Read.me
Includes initial migration of code from STM32 to ESP32 Arduino. There are a few added features in the migration. The most important one is being support for a second myoware sensor was added and a buffer was made for both to store 5 seconds of data. 

The read.me is currently a place holder. It will be updated in the near future