Implemented Current Sensors and Motor Controls Inverted
Added the ability to read the output from the current sensors. Current sensors are not yet calibrated due to the large resistance that comes with the perfboard making them unreliable. A max draw threshold was set at 2V for a place holder value.

Motor controls have been inverted: 
80% duty -> hand
20 % duty -> closed hand
This is to combat the inverting nature of the MOSFET that was recently implemented to drive the servos.